Initial Frontend Setup (#8)
* Init angular project

Run : ng new OmnAIView --skip-git

* Update gitignore, Update App name

Changes:

.gitignore: added the angular folder to gitignore

package.json: Updated the current version to v0.1.0
	      Updated the name to OmnAIView

* Change foldername from OmnAIView->angular-frontend

Problem:

As the project grows the project will not only have a angular frontend
but also a electron app as well as a possible android app. Therefore
the projects should be seperated by their frontend type as they are a
part of OmnAIView.

Changes:

The name for the angular frontend setup is now called: angular frontend
instead of OmnAIView

* Change foldername from OmnAIView->angular-frontend

Problem:

As the project grows the project will not only have a angular frontend
but also a electron app as well as a possible android app. Therefore
the projects should be seperated by their frontend type as they are a
part of OmnAIView.

Changes:

The name for the angular frontend setup is now called: angular frontend
instead of OmnAIView

* README.md Merge

Merged the original README.md with the README.md from the angular
project. Deleted unnessecary information from the angular README.md.
Information now included is how to run and build the project as well as
the version of angular used.

* Update: .gitignore moved one level up

Because the .gitignore in the angular-frontend folder gets ignored by
git, the .gitignore is moved one level up into the root folder of the
project. Deleted the old .gitignore and changed the paths in the
.gitignore to fit the new directory