Init Datasourceinterface description for OmnAIScope-Dataserver (#7)
* Init Datasource Interface with AsyncAPI doc

AsyncAPI is a format to not only document API endpoints but also
websocket endpoints. The Async-API description describes the datasource
interface that is used for OmnAIView. This includes the API as well as
the websocket interface. It also describes in which formats data should
be available from the websocket.

* Update: OmnAIScope-Dataserver update

Problem: The description only fits the OmnAIScope dataservers
not a general dataserver for OmnAIView.

Solution: Therefore the description
was changed to only document the OmnAIScope dataservers.
Currently documented OmnAIScope dataserver version: v0.4.0