Feature: Init Electron setup  (#10)
* Init: Electron with webpack + .gitignore

Ran : npx create-electron-app@latest my-new-app --template=webpack-typescript
Moved everything from the original .gitignore into the main .gitignore.

The electron project as it is is in now way connected
to the angular project currently. Currently this is only a template.

* Update: .gitignore for electron: dist/res folder

Added the dist and res folder that will be created in the electron
folder, to the .gitignore

* Using angular in electron app

As we are not using webpack and electron independend from the angular
application, webpack as well as the html files from the electron project
are no longer used. Instead a main.ts is used that will only render
the angular project. Therefore the angular project is included into the
electron project.

Important!: The tsc version was changed to a higher version because the
older version is no longer supported.

* Update Version and name of Angular project

Name: OmnAIView
Version : 0.1.0

* Update README.md and project name

Updated the project name in the package.json to OmnAIView
Updated the make process to also use tsc when running
npm run build for electron
Updated the README.md with a documenation on `How to run the angular
project`

* Update: Author, email in package.json

Updated the author and email to fit the Auto-Intern GmbH

Author: Auto-Intern GmbH
Email: info@auto-intern.de