Feature/ Start OmnAIScope backend in Autostart (#11)
* Feature: Autostart OmnAIScope BE with frontend

As the OmnAIScope BE can be used locally with the SW, it should be
delivered with the SW as an integrated component. This will make the use
of a local server way easier for the user.

Changes:
- Init omnaibackend.ts which provides an omnaibackendmanager to start the
BE
- Start the backend in the main.ts
- Setup of the backend in the res folder with all it dependencies
- Added the backend as extraResources in the forge.config.ts to be
installed with the local application
- Update of the README.md to document which BE version is used from
  which repo

* Update README.md

Typo

Co-authored-by: Tabea <Tabea.Roethemeyer@rub.de>
Signed-off-by: AKMaily <83161524+AKMaily@users.noreply.github.com>

* Update: Added OmnAIScope BE Files in res folder

Added the OmnAIScope BE Files in the omnai_BE folder to the git
history. Deleted the .gitignore path for those files.

---------

Signed-off-by: AKMaily <83161524+AKMaily@users.noreply.github.com>
Co-authored-by: Tabea <Tabea.Roethemeyer@rub.de>