Feature: Starting Backend on a free port (#13)
* Feature: Starting Backend on a free port

Problem: For v0.5.0 to start the backend with a websocket connection
a port has to be given in the CLI.

Solution: Searching for a free port by testing a free port on the
clients device instead of setting a hardcoded port. Giving the port as
a parameter into spawn when starting the backend.

Changes:
- Implementation of the above behavior
- Changing the app version to 0.2.0 in the package.json as this is a new
  feature

* Update : Searching for port via the OS

Instead of searching for a port by incrementing the port number and
checking if the port is available, the port is now searched by the OS.
This decrements the needed operating time. Added a Typeguard to check if
the received addressinfo is of type Addressinfo.

* Fix: Cleaner Code as suggested

Import AddressInfo from net with * .
Unnecessary 'else' statement deleted.