Feature: RestAPI connection to OmnAIBE alpha
What ?
This commit provides a first code base to get the device list
via Rest API in the angular frontend from the OmnAIScope Backend.

Why ?
To show devices to the user it is important to know which devices are
connected

Important changes :
- Update MiniOmni.exe to v0.5.1 which uses the correct port settings
- Introduction of the backend-handling : Port service and api service
- devicelist component to be able to interact and visualize the received
  API response
- deletion of the standard html file, only the devicelist is visible

Personal Note:
This is an alpha version of the setup to commit the first important
changes needed for a Rest API connection. It only provides code for the
local backend. This should be later on designed more usable with a serverbox
and a more generic code using servers from different IPs and Ports.

Important notes for beginners :
HttpClient need to be listed as a provider in the app.config.ts
otherwise you get weird errors.