PR: Minor changes, adds icon and fixes port selection  (#19)
This PR fixes minor issues. 

Design
The svg graph was about 1px too wide which lead to the generation of a scrollbar, that is not needed.
We created an icon in different formats and added it, to eliminate the default Electron icon (mind: We need a URL that is (best-case) hosted on AI-Gruppe-Servers (see also description in the changed forge.config.ts file)).
Technical
The Backend Server spawns on a random free port. The electron App uses port 8080. That way the Scopes were not recognized. This was fixed by using the already existing getOmnAIScopeBackendPort function and changing the port upon init success. If not, it displays an error in the log and defaults back to :8080.
Changed index.html to index.csr.html. Without it the build just showed a blank page.
Testing:

Build the app as described in README.md, then start and connect a scope. Click "Load Device list" and you will see a scope listed. Otherwise you can also check, that in the dev tools, you see a connection being made insted of an ERR REFUSED.

---------

Signed-off-by: Lukas Adrian Kron <l.kron@lugges.eu>
Co-authored-by: cyblob <erdmann-robert@gmx.de>
Co-authored-by: Robert <116553068+ErdmannRobert@users.noreply.github.com>