Bug/clean backendfix (#24)
* Fix: Electron Build + Connection to OmnAI BE

Enhancement :
- Trying to connect with the Backend works after object instantiation
--> the port information is not available in other classes and only
after object instantiation
--> This could lead to timing problems when trying to reach the BE

Changes:
- backend-port Service which loads and provides the port on which the backend started
- integration of the chosen port in the server url in live-data service
- Init the port information as a signal that is received with the
  initialization of the app

What should work now :
- Running the app with npm start should open the correct html file
You should see the graph and two buttons
- Clicking on the button <load devices> should show you the connected
  OmnAI-Devices underneath.

Current Problems:
- Data from the OmnAIScopes is not shown in the graph
  The dummy data works correctly

* Fix: Receiving data from backend

Problem : Data from the OmnAIScope was not drawn in the graph

Solution: The Backend expects the UUID from the devices that should send
data as a string . The UUID was send as a JSON string which means it had
quotation marks around it. The Backend expects a normal string without
quotation marks. This was fixed by this commit.

* Code Enhancement: Doc and Clean code

Documentation of the Backend-port Service via a comment.
Deletion of not needed code.

* Documentation Update: omnaiBackend.ts

Added comments for the functions of the omnaiscopeBackendManager in the
omnaiBackend.ts
This should document the functions better and make the usage of these
functions easier.

* Documentation Update and Code safety

Changes:
- Doc update for omnaiBackend.ts, backend-port.service.ts and
  app.config.ts because the descriptions where not clear or false
- check if websocket exists before sending UUIDs to the backend
- read out url only once from the signal when trying to get devices from the backend

* Doc: MiniOmni to OmnAIScope Backend

For documentation the name MiniOmni should not be used as this is easily
changeable in the future and does not necessarily lead to the OmnAIScope
backend. Therefor the documentation was updated to reference the
OmnAIScope backend as OmnAIScope Backend directly.