Doc: Add CHANGELOG.md for contributions (#41)
* Doc: Adding a Changelog: Issue #38

Regarding Issue #38 a changelog should be integrated to document the
changes that were/will be made between different versions.
As a default template the KeepAChangelog standard is used.
The initial setup includes all changes that were merged in the master
until this commit.

* Doc: Adding changelog info in contribution.md

As the changelog has to be kept up-to-date contributors should
integrate their changes in the changelog when creating a PR.
Therefor the information about the changelog and the expected changes in
the changelog was added to the CONTRIBUTION.md.

Also added the changes of the CONTRIBUTION.md in the CHANGELOG.