feat(import): support legacy OmnAI CSV import (MVP) (#39)
Add file selecting UI and parser for converting files exported by the old OmnAI software into a format that is 
usable by the new OmnAIView project.
Contains a csv parser that is dependend specifically on the format from the old project.
Feature is still experimental, issues are expected and not all csv files will be parsable. Errors are thrown depending on
when the parser has problems.

Tests: Manual test first exporting data in the old software, then importing into the new one.

Background: Before the OpenSource OmnAI View, a first prototype in C++ was developed and users already used it for recording data. It already had a Save-Feature included so that data could be exported from the old version. To be able to still access old recordings in this OmnAIView project, an import option for this data format is needed.

---
Co-authored-by: Tabea <tabea.roethemeyer@gruppe.ai>
Signed-off-by: C0D3 M4513R <28912031+C0D3-M4513R@users.noreply.github.com>