Doc(API): Deletion of BE Doc intern (#70)
Deletion of the Async API description of the OmnAIScope Dataserver
Backend in this repository.

The Async API description in the repository is a duplicate to the Async
API description in the OmnAIScope DataServer repository.
To have one consistent source of truth the API description of a backend
should be in the repository of the backend itself.